refactoring: move static methods to chared class
