+++ RefactoringEssentials/CSharp/Manipulations.cs
@@ -1,0 +1,56 @@
+﻿using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Formatting;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace RefactoringEssentials.CSharp
+{
+    static class Manipulations
+    {
+        internal static SyntaxNode AddBefore(SyntaxNode root, SyntaxNode loationToAddBefore, SyntaxNode nodeToAdd)
+        {
+            return root.InsertNodesBefore(
+                root.GetCurrentNode(loationToAddBefore)
+                , new List<SyntaxNode> { nodeToAdd });
+        }
+
+        internal static SyntaxNode AddStatementToConstructorBody(SyntaxNode root, ConstructorDeclarationSyntax constructor, StatementSyntax statement)
+        {
+            var body = constructor.Body ?? SyntaxFactory.Block();
+
+            return root.ReplaceNode(root.GetCurrentNode(constructor), constructor.WithBody(
+                    body.WithStatements(SyntaxFactory.List(new[] { statement }.Concat(body.Statements)))
+                ));
+        }
+
+        internal static ExpressionStatementSyntax CreateAssignmentStatement(string leftHandSidePropertyName, string rightHandSidePropertyName)
+        {
+            return SyntaxFactory.ExpressionStatement(
+                SyntaxFactory.AssignmentExpression(
+                    SyntaxKind.SimpleAssignmentExpression,
+                    leftHandSidePropertyName != rightHandSidePropertyName ? (ExpressionSyntax)SyntaxFactory.IdentifierName(leftHandSidePropertyName) : SyntaxFactory.MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression, SyntaxFactory.ThisExpression(), SyntaxFactory.IdentifierName(rightHandSidePropertyName)),
+                    SyntaxFactory.IdentifierName(rightHandSidePropertyName)
+                )
+            ).WithAdditionalAnnotations(Formatter.Annotation);
+        }
+
+        internal static PropertyDeclarationSyntax CreateAutoProperty(TypeSyntax type, string identifier, bool getOnly, SyntaxKind? accessibility)
+        {
+            var accessorDeclList = new SyntaxList<AccessorDeclarationSyntax>()
+            .Add(SyntaxFactory.AccessorDeclaration(SyntaxKind.GetAccessorDeclaration).WithSemicolonToken(SyntaxFactory.Token(SyntaxKind.SemicolonToken)));
+
+            if (getOnly == false)
+                accessorDeclList.Add(SyntaxFactory.AccessorDeclaration(SyntaxKind.SetAccessorDeclaration).WithSemicolonToken(SyntaxFactory.Token(SyntaxKind.SemicolonToken)));
+
+            var newProperty = SyntaxFactory.PropertyDeclaration(type, identifier)
+                .WithAccessorList(SyntaxFactory.AccessorList(accessorDeclList));
+
+            if (accessibility.HasValue)
+                newProperty = newProperty.WithModifiers(SyntaxFactory.TokenList(SyntaxFactory.Token(accessibility.Value)));
+
+            return newProperty.WithAdditionalAnnotations(Formatter.Annotation);
+        }
+    }
+}
